+++ TimeSliceNet/ExampleClasses/GasStation/GasStationClasses.cs
@@ -3,7 +3,6 @@ using TimeSlice;
 
 namespace ExampleClasses.GasStation
 {
-
     /// <summary>
     ///     a car needs gasoline which it gets from a <see cref="GasolinePump" />
     /// </summary>
