+++ TimeSliceNet/ExampleWebApplication/Startup.cs
@@ -12,7 +12,6 @@ namespace ExampleWebApplication
     [ExcludeFromCodeCoverage]
     public class Startup
     {
-
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
